requirements/base.txt
Django==1.5.2
Jinja2==2.7
MarkupSafe==0.18
PIL==1.1.7
Pygments==1.6
South==0.8.1
Sphinx==1.2b1
Unipath==1.0
django-crispy-forms==1.3.2
django-extensions==1.1.1
django-floppyforms==1.1
django-haystack==2.1.0
django-localflavor==1.0
django-model-utils==1.4.0
django-registration==1.0
docutils==0.11
easy-thumbnails==1.3
psycopg2==2.5.1
pyelasticsearch==0.6
python-magic==0.4.3
requests==1.2.3
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2